Simplify bitwise vector test functions
Change `is_all_ones` and `is_all_zeros` to no longer use `std::bitset` (and just use a per-slot test instead).

This also removes a linker error/bug when using these functions that causes the symbols for `std::bitset` to not be exported in a client library under certain conditions.